\name{bison_data}
\alias{bison_data}
\title{Get data from output of bison.}
\usage{
  bison_data(input, ...)
}
\arguments{
  \item{input}{Input data.frame}

  \item{...}{Additional parameters passed on to
  bison_data.data, bison_data.counties, bison_data.states}
}
\description{
  Get data from output of bison.
}

