% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_datause.R
\name{bison_datause}
\alias{bison_datause}
\alias{bison_citation}
\title{Get BISON data use agreement details and examples for how to cite data.}
\usage{
bison_datause()

bison_citation()
}
\description{
Get BISON data use agreement details and examples for how to cite data.
}
\references{
https://bison.usgs.gov/doc/api.jsp#data
}
