\name{bison_map_maker}
\alias{bison_map_maker}
\title{Function to make map form lat/long data}
\usage{
  bison_map_maker(x, geom, jitter, customize)
}
\arguments{
  \item{x}{The input}

  \item{...}{Further parameters passed on to bisonmap
  function}
}
\description{
  Function to make map form lat/long data
}
\keyword{internal}

