% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbison-package.R
\docType{data}
\name{fips}
\alias{fips}
\title{Fips codes for states and counties}
\format{A data frame with 3142 rows and 4 variables:
\describe{
\item{state}{State name}
\item{county}{County name}
\item{fips_state}{State FIPS code}
\item{fips_county}{County FIPS code}
}}
\description{
See
https://www.census.gov/geo/reference/codes/cou.html and
https://www.census.gov/geo/reference/ansi_statetables.html for
more information on FIPS codes
}
\keyword{data}
