\docType{package}
\name{rbison-package}
\alias{rbison}
\alias{rbison-package}
\title{rbison is an interface to the USGS Bison API.}
\description{
To get started, see the vignette in
\code{vignette(package="rbison")}
}
\details{
See [here](http://bison.usgs.ornl.gov/doc/services.jsp) for
API docs for the BISON API.

To cite rbison, do \code{citation(package='rbison')}

Use the following format to cite data retrieved from BISON:

Biodiversity occurrence data published by: (Accessed
through Biodiversity Information Serving our Nation
(BISON), bison.usgs.ornl.gov, YYYY-MM-DD).

For example:

Biodiversity occurrence data published by: Field Museum of
Natural History, Museum of Vertebrate Zoology, University
of Washington Burke Museum, and University of Turku
(Accessed through Biodiversity Information Serving our
Nation (BISON), bison.usgs.ornl.gov, 2013-04-22).
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

